install node exporter (#42)
Changes:

Instead of using `ansible.builtin.unarchive` to pull and inflate the
node exporter binary, which is skipped when the destination folder
already exists, add a task to download the binary to a temporary
location and a task for copying into place. An updated version of node
exporter was not being installed on clients where an old version already
existed.