Run `molecule test` on `install_xnat` playbook (#62)
- run `molecule test` on `mirsg.infrastructure.install_xnat` playbook.
This is to check that the playbook can be run twice without failure, and
for idempotency checks
- update the `centos7_xnat` scenario to set the `/etc/resolv.conf` for
the containers. This defines nameservers for the containers, which for
some reason was necessary if we also use firewalld (otherwise the
containers can't reach the internet after firewalld is started)
- update the `rocky9_xnat` scenario to set the `/etc/resolv.conf` for
the containers. For some reason, this is not necessary locally but is
required for the tests to pass in CI