Update container service (#130)
Adds support for a container service to serve multiple XNAT clients.

The main change is modifying the provision role to allow checking of
multiple mount points. Previously only a single mount point was
supported, but a container server may have a separate mount for each
server it supports.

Multiple mount points can be configured by defining the list
`external_storage_mounts` in the host variables for the container
server. If not set, the default will fall back to a list containing the
single entry `external_storage_drive`, or an empty list if this is not
set. This preserves existing behaviour.

The container service client role will now configure the XNAT container
service "Docker image hosts" (aka hubs) from which containers can be
pulled to run on XNAT. These can be defined in the Ansible variables.

In addition, I've added two optional changes to facilitate local
testing:
- Allow optional use of the GitHub package registry instead of JFrog, by
permitting maven_artifact to optionally use username/password
authentication instead of an authentication header. This preserves
existing behaviour but permits switching to a package registry with a
different authentication mechanism if appropriately configured in the
ansible variables
- Install firewalld package as part of the firewall role if it is not
present